+++ src/OptionsStrategy.php
@@ -33,26 +33,26 @@ function portfolioId($p) {
 	return "".$p->call_put_flag."_".$p->X."_".$p->r."_".$p->v;
 }
 
-function loss($S,$T,$mult=1) {
+function loss($S,$t,$mult=1) {
 // $S: underlying value
-// $T: time to maturity
+// $t: physical time
 // $mult: factor with which to multiply the loss... useful for this->margin
 
 	if(!is_numeric($mult)) throw new Exception("mult should be numeric");
-	return $this->loss2($this->loss1($S,$T),$mult,$S,$T);
+	return $this->loss2($this->loss1($S,$t),$mult,$S,$t);
 }
 
-function loss1($S,$T) {
+function loss1($S,$t) {
 // sub-function of loss
 	$vals=array();
 	foreach($this->portfolio as $id=>$p) {
-		$v=$p["o"]->value($S,$T);
+		$v=$p["o"]->value($S,$t);
 		if(is_array($v)) {
 			$q=$p["q"];
 			// 1st dimension: S
 			$v=array_map(function($vi) use($q) {
 				if(is_array($vi)) {
-					// 2nd dimension: T
+					// 2nd dimension: t
 					return array_map(function($vii) use($q) { return $vii*$q; },$vi);
 				} else {
 					return $vi*$q;
@@ -62,7 +62,7 @@ function loss1($S,$T) {
 			$v=$v*$p["q"];
 		}
 
-		// swap dimensions of options with (S,T)
+		// swap dimensions of options with (S,t)
 		if(is_array($v)) {
 			foreach($v as $ki=>$vi) {
 				if(!array_key_exists($ki,$vals)) $vals[$ki]=array();
@@ -83,13 +83,13 @@ function loss1($S,$T) {
 	return $vals;
 }
 
-function loss2($vals,$mult,$S,$T) {
+function loss2($vals,$mult,$S,$t) {
 // sub-function of loss
-	if(!is_array($S)&&!is_array($T)) {
+	if(!is_array($S)&&!is_array($t)) {
 		$vals = array_sum($vals);
 		$vals = min($vals,0)*$mult;
 	} else {
-		if(is_array($S)&&is_array($T)) {
+		if(is_array($S)&&is_array($t)) {
 			foreach($vals as $ki=>$vi) {
 				foreach($vi as $kii=>$vii) {
 					$vii = array_sum($vii);
@@ -98,7 +98,7 @@ function loss2($vals,$mult,$S,$T) {
 				}
 			}
 		} else {
-			if(!is_array($S)||!is_array($T)) {
+			if(!is_array($S)||!is_array($t)) {
 				foreach($vals as $ki=>$vi) {
 					$vi = array_sum($vi);
 					$vi = min($vi,0)*$mult;
@@ -112,17 +112,17 @@ function loss2($vals,$mult,$S,$T) {
 	return $vals;
 }
 
-function margin($S,$T) {
-	$m=$this->loss($S,$T,-1);
+function margin($S,$t) {
+	$m=$this->loss($S,$t,-1);
 
-	if(!is_array($S)&&!is_array($T)) {
+	if(!is_array($S)&&!is_array($t)) {
 		if(count($this->portfolio)==1) return $m; // this would be numeric in this case
 		return $m;
 	} else {
-		if(is_array($S)&&is_array($T)) {
+		if(is_array($S)&&is_array($t)) {
 			return max(max($m));
 		} else {
-			if(!is_array($S)||!is_array($T)) {
+			if(!is_array($S)||!is_array($t)) {
 				return max($m);
 			} else {
 				throw new Exception("WTF");