+++ tests/OptionsStrategyTest.php
@@ -7,7 +7,7 @@ class OptionsStrategyTest extends PHPUnit_Framework_TestCase
 {
 
     public function testLongCall() {
-	$bs1=new OptionsStrategy(array(array("q"=>1,"o"=>new BlackScholes('c',80,0,0.25))));
+	$bs1=new OptionsStrategy(array(array("q"=>1,"o"=>new BlackScholes('c',80,0,0.25,0))));
 	$this->assertTrue($bs1->loss(120,0)==0);
 	$this->assertTrue($bs1->loss(100,0)==0);
 	$this->assertTrue($bs1->loss(80,0)==0);
@@ -15,7 +15,7 @@ class OptionsStrategyTest extends PHPUnit_Framework_TestCase
     }
 
     public function testShortCall() {
-	$bs1=new OptionsStrategy(array(array("q"=>-1,"o"=>new BlackScholes('c',80,0,0.25))));
+	$bs1=new OptionsStrategy(array(array("q"=>-1,"o"=>new BlackScholes('c',80,0,0.25,0))));
 	$this->assertTrue($bs1->loss(120,0)==-40);
 	$this->assertTrue($bs1->loss(100,0)==-20);
 	$this->assertTrue($bs1->loss(80,0)==0);
@@ -24,14 +24,18 @@ class OptionsStrategyTest extends PHPUnit_Framework_TestCase
 
     public function testDupes() {
 	try {
-		$bs1=new OptionsStrategy(array(array("q"=>2,"o"=>new BlackScholes('c',80,0,0.25))));
+		$bs1=new OptionsStrategy(array(
+			array("q"=>2,"o"=>new BlackScholes('c',80,0,0.25,0)),
+			array("q"=>2,"o"=>new BlackScholes('c',80,0,0.25,0))
+		));
+		$this->assertTrue(false); // should not get here
 	} catch(Exception $e) {
 		$this->assertTrue($e->getMessage()=="Duplicate options in portfolio detected");
 	}
     }
 
     public function testLong2xCall() {
-	$bs2=new OptionsStrategy(array(array("q"=>2,"o"=>new BlackScholes('c',80,0,0.25))));
+	$bs2=new OptionsStrategy(array(array("q"=>2,"o"=>new BlackScholes('c',80,0,0.25,0))));
 	$this->assertTrue($bs2->loss(120,0)==0);
 	$this->assertTrue($bs2->loss(100,0)==0);
 	$this->assertTrue($bs2->loss(80,0)==0);
@@ -39,7 +43,7 @@ class OptionsStrategyTest extends PHPUnit_Framework_TestCase
     }
 
     public function testShort2xCall() {
-	$bs1=new OptionsStrategy(array(array("q"=>-2,"o"=>new BlackScholes('c',80,0,0.25))));
+	$bs1=new OptionsStrategy(array(array("q"=>-2,"o"=>new BlackScholes('c',80,0,0.25,0))));
 	$this->assertTrue($bs1->loss(120,0)==-80);
 	$this->assertTrue($bs1->loss(100,0)==-40);
 	$this->assertTrue($bs1->loss(80,0)==0);
@@ -47,7 +51,7 @@ class OptionsStrategyTest extends PHPUnit_Framework_TestCase
     }
 
     public function testQtyZero() {
-	$bs2=new OptionsStrategy(array(array("q"=>0,"o"=>new BlackScholes('c',80,0,0.25))));
+	$bs2=new OptionsStrategy(array(array("q"=>0,"o"=>new BlackScholes('c',80,0,0.25,0))));
 	$this->assertTrue($bs2->loss(120,0)==0);
 	$this->assertTrue($bs2->loss(100,0)==0);
 	$this->assertTrue($bs2->loss(80,0)==0);
@@ -55,7 +59,7 @@ class OptionsStrategyTest extends PHPUnit_Framework_TestCase
     }
 
     public function testLongPut() {
-	$bs1=new OptionsStrategy(array(array("q"=>1,"o"=>new BlackScholes('p',80,0,0.25))));
+	$bs1=new OptionsStrategy(array(array("q"=>1,"o"=>new BlackScholes('p',80,0,0.25,0))));
 	$this->assertTrue($bs1->loss(120,0)==0);
 	$this->assertTrue($bs1->loss(100,0)==0);
 	$this->assertTrue($bs1->loss(80,0)==0);
@@ -63,7 +67,7 @@ class OptionsStrategyTest extends PHPUnit_Framework_TestCase
     }
 
     public function testShortPut() {
-	$bs1=new OptionsStrategy(array(array("q"=>-1,"o"=>new BlackScholes('p',80,0,0.25))));
+	$bs1=new OptionsStrategy(array(array("q"=>-1,"o"=>new BlackScholes('p',80,0,0.25,0))));
 	$this->assertTrue($bs1->loss(120,0)==0);
 	$this->assertTrue($bs1->loss(100,0)==0);
 	$this->assertTrue($bs1->loss(80,0)==0);
@@ -73,40 +77,40 @@ class OptionsStrategyTest extends PHPUnit_Framework_TestCase
     public function testMarginNumeric() {
 	// test that margin is always numeric
 
-	$bs1=new OptionsStrategy(array(array("q"=>1,"o"=>new BlackScholes('p',80,0,0.25))));
+	$bs1=new OptionsStrategy(array(array("q"=>1,"o"=>new BlackScholes('p',80,0,0.25,0))));
 	$x=$bs1->margin(60,0);
 	$this->assertTrue(is_numeric($x));
 
-	$bs1=new OptionsStrategy(array(array("q"=>1,"o"=>new BlackScholes('p',80,0,0.25))));
+	$bs1=new OptionsStrategy(array(array("q"=>1,"o"=>new BlackScholes('p',80,0,0.25,0))));
 	$x=$bs1->margin(array(60,80,100,120),0);
 	$this->assertTrue(is_numeric($x));
 
-	$bs1=new OptionsStrategy(array(array("q"=>1,"o"=>new BlackScholes('p',80,0,0.25))));
+	$bs1=new OptionsStrategy(array(array("q"=>1,"o"=>new BlackScholes('p',80,0,0.25,0.2))));
 	$x=$bs1->margin(array(60,80,100,120),array(0,0.1,0.2));
 	$this->assertTrue(is_numeric($x));
 
 	$bs1=new OptionsStrategy(array(
-		array("q"=> 1,"o"=>new BlackScholes('p',80,0,0.25)),
-		array("q"=>-1,"o"=>new BlackScholes('p',90,0,0.25))
+		array("q"=> 1,"o"=>new BlackScholes('p',80,0,0.25,0)),
+		array("q"=>-1,"o"=>new BlackScholes('p',90,0,0.25,0))
 	));
 	$x=$bs1->margin(60,0);
 	$this->assertTrue(is_numeric($x));
     }
 
     public function testVector() {
-	$bs1=new OptionsStrategy(array(array("q"=>1,"o"=>new BlackScholes('p',80,0,0.25))));
+	$bs1=new OptionsStrategy(array(array("q"=>1,"o"=>new BlackScholes('p',80,0,0.25,0))));
 	$x=$bs1->loss(array(60,80,100,120),0);
 	$this->assertTrue(count($x)==4);
     }
 
     public function testShortCallVector() {
-	$bs1=new OptionsStrategy(array(array("q"=>-1,"o"=>new BlackScholes('c',80,0,0.25))));
+	$bs1=new OptionsStrategy(array(array("q"=>-1,"o"=>new BlackScholes('c',80,0,0.25,0))));
 	$x=$bs1->margin(array(60,80,100,120),0);
 	$this->assertTrue($x==40);
     }
 
     public function testTensor() {
-	$bs1=new OptionsStrategy(array(array("q"=>1,"o"=>new BlackScholes('p',80,0,0.25))));
+	$bs1=new OptionsStrategy(array(array("q"=>1,"o"=>new BlackScholes('p',80,0,0.25,0.2))));
 	$x3=$bs1->loss(array(60,80,100,120),array(0,0.1,0.2));
 	$this->assertTrue(count($x3)==4);
 	foreach($x3 as $k=>$x3i) {
@@ -115,15 +119,15 @@ class OptionsStrategyTest extends PHPUnit_Framework_TestCase
     }
 
     public function testShortCallTensor() {
-	$bs1=new OptionsStrategy(array(array("q"=>-1,"o"=>new BlackScholes('c',80,0,0.25))));
+	$bs1=new OptionsStrategy(array(array("q"=>-1,"o"=>new BlackScholes('c',80,0,0.25,0.2))));
 	$x=$bs1->margin(array(60,80,100,120),array(0,0.1,0.2));
 	$this->assertTrue($x>40 && $x<41);
     }
 
     public function testStraddleLong() {
 	$bs1=new OptionsStrategy(array(
-		array("q"=>1,"o"=>new BlackScholes('c',80,0,0.25)),
-		array("q"=>1,"o"=>new BlackScholes('p',80,0,0.25))
+		array("q"=>1,"o"=>new BlackScholes('c',80,0,0.25,0.2)),
+		array("q"=>1,"o"=>new BlackScholes('p',80,0,0.25,0.2))
 	));
 	$x=$bs1->margin(array(60,80,100,120),array(0,0.1,0.2));
 	$this->assertTrue($x==0);
@@ -131,8 +135,8 @@ class OptionsStrategyTest extends PHPUnit_Framework_TestCase
 
     public function testStraddleShort() {
 	$bs1=new OptionsStrategy(array(
-		array("q"=>-1,"o"=>new BlackScholes('c',80,0,0.25)),
-		array("q"=>-1,"o"=>new BlackScholes('p',80,0,0.25))
+		array("q"=>-1,"o"=>new BlackScholes('c',80,0,0.25,0.2)),
+		array("q"=>-1,"o"=>new BlackScholes('p',80,0,0.25,0.2))
 	));
 	$x=$bs1->margin(array(60,80,100,120),array(0,0.1,0.2));
 	$this->assertTrue($x>40&&$x<41);
@@ -140,8 +144,8 @@ class OptionsStrategyTest extends PHPUnit_Framework_TestCase
 
     public function testSpreadLong() {
 	$bs1=new OptionsStrategy(array(
-		array("q"=> 1,"o"=>new BlackScholes('c',80,0,0.25)),
-		array("q"=>-1,"o"=>new BlackScholes('c',90,0,0.25))
+		array("q"=> 1,"o"=>new BlackScholes('c',80,0,0.25,0.2)),
+		array("q"=>-1,"o"=>new BlackScholes('c',90,0,0.25,0.2))
 	));
 	$x=$bs1->margin(array(60,80,100,120),array(0,0.1,0.2));
 	$this->assertTrue($x==0);
@@ -149,8 +153,8 @@ class OptionsStrategyTest extends PHPUnit_Framework_TestCase
 
     public function testSpreadShort() {
 	$bs1=new OptionsStrategy(array(
-		array("q"=>-1,"o"=>new BlackScholes('c',80,0,0.25)),
-		array("q"=> 1,"o"=>new BlackScholes('c',90,0,0.25))
+		array("q"=>-1,"o"=>new BlackScholes('c',80,0,0.25,0.2)),
+		array("q"=> 1,"o"=>new BlackScholes('c',90,0,0.25,0.2))
 	));
 	$x=$bs1->margin(array(60,80,100,120),array(0,0.1,0.2));
 	$this->assertTrue($x==10);
@@ -158,10 +162,10 @@ class OptionsStrategyTest extends PHPUnit_Framework_TestCase
 
     public function testCondorLong() {
 	$bs1=new OptionsStrategy(array(
-		array("q"=>-1,"o"=>new BlackScholes('c',70,0,0.25)),
-		array("q"=> 1,"o"=>new BlackScholes('c',80,0,0.25)),
-		array("q"=> 1,"o"=>new BlackScholes('c',90,0,0.25)),
-		array("q"=>-1,"o"=>new BlackScholes('c',100,0,0.25))
+		array("q"=>-1,"o"=>new BlackScholes('c',70,0,0.25,0.2)),
+		array("q"=> 1,"o"=>new BlackScholes('c',80,0,0.25,0.2)),
+		array("q"=> 1,"o"=>new BlackScholes('c',90,0,0.25,0.2)),
+		array("q"=>-1,"o"=>new BlackScholes('c',100,0,0.25,0.2))
 	));
 	$x=$bs1->margin(array(60,80,100,120),array(0,0.1,0.2));
 	$this->assertTrue($x==10);
@@ -169,10 +173,10 @@ class OptionsStrategyTest extends PHPUnit_Framework_TestCase
 
     public function testCondorShort() {
 	$bs1=new OptionsStrategy(array(
-		array("q"=> 1,"o"=>new BlackScholes('c',70,0,0.25)),
-		array("q"=>-1,"o"=>new BlackScholes('c',80,0,0.25)),
-		array("q"=>-1,"o"=>new BlackScholes('c',90,0,0.25)),
-		array("q"=> 1,"o"=>new BlackScholes('c',100,0,0.25))
+		array("q"=> 1,"o"=>new BlackScholes('c',70,0,0.25,0.2)),
+		array("q"=>-1,"o"=>new BlackScholes('c',80,0,0.25,0.2)),
+		array("q"=>-1,"o"=>new BlackScholes('c',90,0,0.25,0.2)),
+		array("q"=> 1,"o"=>new BlackScholes('c',100,0,0.25,0.2))
 	));
 	$x=$bs1->margin(array(60,80,100,120),array(0,0.1,0.2));
 	$this->assertTrue($x==0);
@@ -180,7 +184,7 @@ class OptionsStrategyTest extends PHPUnit_Framework_TestCase
 
     public function testNotBlackScholes() {
 	try {
-		$bs1=new OptionsStrategy(array(array("q"=>2,"o"=>array(1,2,3,4))));
+		$bs1=new OptionsStrategy(array(array("q"=>2,"o"=>array(1,2,3,4,5))));
 	} catch(Exception $e) {
 		$this->assertTrue($e->getMessage()=="Portfolio should be array of BlackScholes objects");
 	}
@@ -189,9 +193,10 @@ class OptionsStrategyTest extends PHPUnit_Framework_TestCase
     public function testQtyNotNumeric() {
 	try {
 		$bs1=new OptionsStrategy(array(
-			array("q"=>-1,"o"=>new BlackScholes('c',80,0,0.25)),
-			array("q"=> "1","o"=>new BlackScholes('c',90,0,0.25))
+			array("q"=>-1,"o"=>new BlackScholes('c',80,0,0.25,0)),
+			array("q"=> "1","o"=>new BlackScholes('c',90,0,0.25,0))
 		));
+		$this->assertTrue(false); // should not get here
 	} catch(Exception $e) {
 		$this->assertTrue($e->getMessage()=="quantity should be numeric integer");
 	}