+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 2015-06-15
 * added php api backend with hook to calculate margin (along with tests of course)
+* api: Tmt to be per option
+* javascript: add physical time
+* api: round 2nd decimal
+* index: add link to github
 
 2015-06-12
 * can set time to maturity for each instrument (allows calendar spreads)